Viewlet Refactor (#85566)
* wip 2

* use viewlet stubs

* fix layering issue

* fix saveState and contextmenus

* move context menu logic into panecomposite

* revert tasks.json

* tasks again

* add showHeaderInTitleWhenSingleView values

* patch tests

* some cleanup

* more cleanup

* start view

* merging conflicts

* fix title and dnd

* adding missing api surface to pane composite

* additional merge resolution
