+++ src/vs/workbench/browser/parts/views/viewsViewlet.ts
@@ -4,330 +4,30 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as DOM from 'vs/base/browser/dom';
-import { dispose, IDisposable, combinedDisposable, toDisposable } from 'vs/base/common/lifecycle';
 import { IAction } from 'vs/base/common/actions';
 import { Separator } from 'vs/base/browser/ui/actionbar/actionbar';
-import { firstIndex } from 'vs/base/common/arrays';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IContextMenuService } from 'vs/platform/contextview/browser/contextView';
-import { IViewDescriptor, IViewsViewlet, IViewContainersRegistry, Extensions as ViewContainerExtensions, IView } from 'vs/workbench/common/views';
+import { IViewDescriptor } from 'vs/workbench/common/views';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { IStorageService, StorageScope } from 'vs/platform/storage/common/storage';
+import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { IContextKeyService } from 'vs/platform/contextkey/common/contextkey';
-import { StandardMouseEvent } from 'vs/base/browser/mouseEvent';
-import { PaneViewlet, ViewletPane, IViewletPaneOptions } from 'vs/workbench/browser/parts/views/paneViewlet';
-import { DefaultPaneDndController } from 'vs/base/browser/ui/splitview/paneview';
+import { ViewPaneContainer, ViewPane, IViewPaneOptions } from 'vs/workbench/browser/parts/views/viewPaneContainer';
 import { WorkbenchTree, IListService } from 'vs/platform/list/browser/listService';
 import { IWorkbenchThemeService, IFileIconTheme } from 'vs/workbench/services/themes/common/workbenchThemeService';
 import { ITreeConfiguration, ITreeOptions } from 'vs/base/parts/tree/browser/tree';
 import { Event } from 'vs/base/common/event';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
-import { localize } from 'vs/nls';
-import { IAddedViewDescriptorRef, IViewDescriptorRef, PersistentContributableViewsModel } from 'vs/workbench/browser/parts/views/views';
-import { Registry } from 'vs/platform/registry/common/platform';
-import { MementoObject } from 'vs/workbench/common/memento';
+import { IAddedViewDescriptorRef } from 'vs/workbench/browser/parts/views/views';
 
-export interface IViewletViewOptions extends IViewletPaneOptions {
-	viewletState: MementoObject;
+export interface IViewletViewOptions extends IViewPaneOptions {
 }
 
-export abstract class ViewContainerViewlet extends PaneViewlet implements IViewsViewlet {
-
-	private readonly viewletState: MementoObject;
-	private didLayout = false;
-	private dimension: DOM.Dimension | undefined;
-	private areExtensionsReady: boolean = false;
-
-	private readonly visibleViewsCountFromCache: number | undefined;
-	private readonly visibleViewsStorageId: string;
-	protected readonly viewsModel: PersistentContributableViewsModel;
-	private viewDisposables: IDisposable[] = [];
-
-	constructor(
-		id: string,
-		viewletStateStorageId: string,
-		showHeaderInTitleWhenSingleView: boolean,
-		@IConfigurationService configurationService: IConfigurationService,
-		@IWorkbenchLayoutService layoutService: IWorkbenchLayoutService,
-		@ITelemetryService telemetryService: ITelemetryService,
-		@IStorageService protected storageService: IStorageService,
-		@IInstantiationService protected instantiationService: IInstantiationService,
-		@IThemeService themeService: IThemeService,
-		@IContextMenuService protected contextMenuService: IContextMenuService,
-		@IExtensionService protected extensionService: IExtensionService,
-		@IWorkspaceContextService protected contextService: IWorkspaceContextService
-	) {
-		super(id, { showHeaderInTitleWhenSingleView, dnd: new DefaultPaneDndController() }, configurationService, layoutService, contextMenuService, telemetryService, themeService, storageService);
-
-		const container = Registry.as<IViewContainersRegistry>(ViewContainerExtensions.ViewContainersRegistry).get(id);
-		if (!container) {
-			throw new Error('Could not find container');
-		}
-		this.viewsModel = this._register(this.instantiationService.createInstance(PersistentContributableViewsModel, container, viewletStateStorageId));
-		this.viewletState = this.getMemento(StorageScope.WORKSPACE);
-
-		this.visibleViewsStorageId = `${id}.numberOfVisibleViews`;
-		this.visibleViewsCountFromCache = this.storageService.getNumber(this.visibleViewsStorageId, StorageScope.WORKSPACE, undefined);
-		this._register(toDisposable(() => this.viewDisposables = dispose(this.viewDisposables)));
-	}
-
-	create(parent: HTMLElement): void {
-		super.create(parent);
-		this._register(this.onDidSashChange(() => this.saveViewSizes()));
-		this.viewsModel.onDidAdd(added => this.onDidAddViews(added));
-		this.viewsModel.onDidRemove(removed => this.onDidRemoveViews(removed));
-		const addedViews: IAddedViewDescriptorRef[] = this.viewsModel.visibleViewDescriptors.map((viewDescriptor, index) => {
-			const size = this.viewsModel.getSize(viewDescriptor.id);
-			const collapsed = this.viewsModel.isCollapsed(viewDescriptor.id);
-			return ({ viewDescriptor, index, size, collapsed });
-		});
-		if (addedViews.length) {
-			this.onDidAddViews(addedViews);
-		}
-
-		// Update headers after and title contributed views after available, since we read from cache in the beginning to know if the viewlet has single view or not. Ref #29609
-		this.extensionService.whenInstalledExtensionsRegistered().then(() => {
-			this.areExtensionsReady = true;
-			if (this.panes.length) {
-				this.updateTitleArea();
-				this.updateViewHeaders();
-			}
-		});
-
-		this.focus();
-	}
-
-	getContextMenuActions(viewDescriptor?: IViewDescriptor): IAction[] {
-		const result: IAction[] = [];
-		if (viewDescriptor) {
-			result.push(<IAction>{
-				id: `${viewDescriptor.id}.removeView`,
-				label: localize('hideView', "Hide"),
-				enabled: viewDescriptor.canToggleVisibility,
-				run: () => this.toggleViewVisibility(viewDescriptor.id)
-			});
-		}
-		const viewToggleActions = this.viewsModel.viewDescriptors.map(viewDescriptor => (<IAction>{
-			id: `${viewDescriptor.id}.toggleVisibility`,
-			label: viewDescriptor.name,
-			checked: this.viewsModel.isVisible(viewDescriptor.id),
-			enabled: viewDescriptor.canToggleVisibility,
-			run: () => this.toggleViewVisibility(viewDescriptor.id)
-		}));
-
-		if (result.length && viewToggleActions.length) {
-			result.push(new Separator());
-		}
-		result.push(...viewToggleActions);
-
-		const parentActions = this.getViewletContextMenuActions();
-		if (result.length && parentActions.length) {
-			result.push(new Separator());
-		}
-		result.push(...parentActions);
-
-		return result;
-	}
-
-	protected getViewletContextMenuActions() {
-		return super.getContextMenuActions();
-	}
-
-	setVisible(visible: boolean): void {
-		super.setVisible(visible);
-		this.panes.filter(view => view.isVisible() !== visible)
-			.map((view) => view.setVisible(visible));
-	}
-
-	openView(id: string, focus?: boolean): IView {
-		if (focus) {
-			this.focus();
-		}
-		let view = this.getView(id);
-		if (!view) {
-			this.toggleViewVisibility(id);
-		}
-		view = this.getView(id)!;
-		view.setExpanded(true);
-		if (focus) {
-			view.focus();
-		}
-		return view;
-	}
-
-	movePane(from: ViewletPane, to: ViewletPane): void {
-		const fromIndex = firstIndex(this.panes, pane => pane === from);
-		const toIndex = firstIndex(this.panes, pane => pane === to);
-		const fromViewDescriptor = this.viewsModel.visibleViewDescriptors[fromIndex];
-		const toViewDescriptor = this.viewsModel.visibleViewDescriptors[toIndex];
-
-		super.movePane(from, to);
-		this.viewsModel.move(fromViewDescriptor.id, toViewDescriptor.id);
-	}
-
-	layout(dimension: DOM.Dimension): void {
-		super.layout(dimension);
-		this.dimension = dimension;
-		if (this.didLayout) {
-			this.saveViewSizes();
-		} else {
-			this.didLayout = true;
-			this.restoreViewSizes();
-		}
-	}
-
-	getOptimalWidth(): number {
-		const additionalMargin = 16;
-		const optimalWidth = Math.max(...this.panes.map(view => view.getOptimalWidth() || 0));
-		return optimalWidth + additionalMargin;
-	}
-
-	protected isSingleView(): boolean {
-		if (!super.isSingleView()) {
-			return false;
-		}
-		if (!this.areExtensionsReady) {
-			if (this.visibleViewsCountFromCache === undefined) {
-				return false;
-			}
-			// Check in cache so that view do not jump. See #29609
-			return this.visibleViewsCountFromCache === 1;
-		}
-		return true;
-	}
-
-	protected createView(viewDescriptor: IViewDescriptor, options: IViewletViewOptions): ViewletPane {
-		return (this.instantiationService as any).createInstance(viewDescriptor.ctorDescriptor.ctor, ...(viewDescriptor.ctorDescriptor.arguments || []), options) as ViewletPane;
-	}
-
-	protected getView(id: string): ViewletPane | undefined {
-		return this.panes.filter(view => view.id === id)[0];
-	}
-
-	protected onDidAddViews(added: IAddedViewDescriptorRef[]): ViewletPane[] {
-		const panesToAdd: { pane: ViewletPane, size: number, index: number }[] = [];
-		for (const { viewDescriptor, collapsed, index, size } of added) {
-			const pane = this.createView(viewDescriptor,
-				{
-					id: viewDescriptor.id,
-					title: viewDescriptor.name,
-					actionRunner: this.getActionRunner(),
-					expanded: !collapsed,
-					viewletState: this.viewletState
-				});
-			pane.render();
-			const contextMenuDisposable = DOM.addDisposableListener(pane.draggableElement, 'contextmenu', e => {
-				e.stopPropagation();
-				e.preventDefault();
-				this.onContextMenu(new StandardMouseEvent(e), viewDescriptor);
-			});
-
-			const collapseDisposable = Event.latch(Event.map(pane.onDidChange, () => !pane.isExpanded()))(collapsed => {
-				this.viewsModel.setCollapsed(viewDescriptor.id, collapsed);
-			});
-
-			this.viewDisposables.splice(index, 0, combinedDisposable(contextMenuDisposable, collapseDisposable));
-			panesToAdd.push({ pane, size: size || pane.minimumSize, index });
-		}
-
-		this.addPanes(panesToAdd);
-		this.restoreViewSizes();
-
-		const panes: ViewletPane[] = [];
-		for (const { pane } of panesToAdd) {
-			pane.setVisible(this.isVisible());
-			panes.push(pane);
-		}
-		return panes;
-	}
-
-	private onDidRemoveViews(removed: IViewDescriptorRef[]): void {
-		removed = removed.sort((a, b) => b.index - a.index);
-		const panesToRemove: ViewletPane[] = [];
-		for (const { index } of removed) {
-			const [disposable] = this.viewDisposables.splice(index, 1);
-			disposable.dispose();
-			panesToRemove.push(this.panes[index]);
-		}
-		this.removePanes(panesToRemove);
-		dispose(panesToRemove);
-	}
-
-	private onContextMenu(event: StandardMouseEvent, viewDescriptor: IViewDescriptor): void {
-		event.stopPropagation();
-		event.preventDefault();
-
-		const actions: IAction[] = this.getContextMenuActions(viewDescriptor);
-
-		let anchor: { x: number, y: number } = { x: event.posx, y: event.posy };
-		this.contextMenuService.showContextMenu({
-			getAnchor: () => anchor,
-			getActions: () => actions
-		});
-	}
-
-	protected toggleViewVisibility(viewId: string): void {
-		const visible = !this.viewsModel.isVisible(viewId);
-		type ViewsToggleVisibilityClassification = {
-			viewId: { classification: 'SystemMetaData', purpose: 'FeatureInsight' };
-			visible: { classification: 'SystemMetaData', purpose: 'FeatureInsight' };
-		};
-		this.telemetryService.publicLog2<{ viewId: String, visible: boolean }, ViewsToggleVisibilityClassification>('views.toggleVisibility', { viewId, visible });
-		this.viewsModel.setVisible(viewId, visible);
-	}
-
-	private saveViewSizes(): void {
-		// Save size only when the layout has happened
-		if (this.didLayout) {
-			for (const view of this.panes) {
-				this.viewsModel.setSize(view.id, this.getPaneSize(view));
-			}
-		}
-	}
-
-	private restoreViewSizes(): void {
-		// Restore sizes only when the layout has happened
-		if (this.didLayout) {
-			let initialSizes;
-			for (let i = 0; i < this.viewsModel.visibleViewDescriptors.length; i++) {
-				const pane = this.panes[i];
-				const viewDescriptor = this.viewsModel.visibleViewDescriptors[i];
-				const size = this.viewsModel.getSize(viewDescriptor.id);
-
-				if (typeof size === 'number') {
-					this.resizePane(pane, size);
-				} else {
-					initialSizes = initialSizes ? initialSizes : this.computeInitialSizes();
-					this.resizePane(pane, initialSizes.get(pane.id) || 200);
-				}
-			}
-		}
-	}
-
-	private computeInitialSizes(): Map<string, number> {
-		const sizes: Map<string, number> = new Map<string, number>();
-		if (this.dimension) {
-			const totalWeight = this.viewsModel.visibleViewDescriptors.reduce((totalWeight, { weight }) => totalWeight + (weight || 20), 0);
-			for (const viewDescriptor of this.viewsModel.visibleViewDescriptors) {
-				sizes.set(viewDescriptor.id, this.dimension.height * (viewDescriptor.weight || 20) / totalWeight);
-			}
-		}
-		return sizes;
-	}
-
-	protected saveState(): void {
-		this.panes.forEach((view) => view.saveState());
-		this.storageService.store(this.visibleViewsStorageId, this.length, StorageScope.WORKSPACE);
-
-		super.saveState();
-	}
-}
-
-export abstract class FilterViewContainerViewlet extends ViewContainerViewlet {
+export abstract class FilterViewPaneContainer extends ViewPaneContainer {
 	private constantViewDescriptors: Map<string, IViewDescriptor> = new Map();
 	private allViews: Map<string, Map<string, IViewDescriptor>> = new Map();
 	private filterValue: string | undefined;
@@ -345,7 +45,8 @@ export abstract class FilterViewContainerViewlet extends ViewContainerViewlet {
 		@IExtensionService extensionService: IExtensionService,
 		@IWorkspaceContextService contextService: IWorkspaceContextService
 	) {
-		super(viewletId, `${viewletId}.state`, false, configurationService, layoutService, telemetryService, storageService, instantiationService, themeService, contextMenuService, extensionService, contextService);
+
+		super(viewletId, `${viewletId}.state`, { showHeaderInTitleWhenSingleView: false }, instantiationService, configurationService, layoutService, contextMenuService, telemetryService, extensionService, themeService, storageService, contextService);
 		this._register(onDidChangeFilterValue(newFilterValue => {
 			this.filterValue = newFilterValue;
 			this.onFilterChanged(newFilterValue);
@@ -397,7 +98,7 @@ export abstract class FilterViewContainerViewlet extends ViewContainerViewlet {
 		}));
 
 		result.push(...viewToggleActions);
-		const parentActions = this.getViewletContextMenuActions();
+		const parentActions = super.getContextMenuActions();
 		if (viewToggleActions.length && parentActions.length) {
 			result.push(new Separator());
 		}
@@ -423,8 +124,8 @@ export abstract class FilterViewContainerViewlet extends ViewContainerViewlet {
 		return views;
 	}
 
-	onDidAddViews(added: IAddedViewDescriptorRef[]): ViewletPane[] {
-		const panes: ViewletPane[] = super.onDidAddViews(added);
+	onDidAddViews(added: IAddedViewDescriptorRef[]): ViewPane[] {
+		const panes: ViewPane[] = super.onDidAddViews(added);
 		for (let i = 0; i < added.length; i++) {
 			if (this.constantViewDescriptors.has(added[i].viewDescriptor.id)) {
 				panes[i].setExpanded(false);
