+++ src/vs/workbench/common/viewlet.ts
@@ -3,14 +3,14 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IComposite } from 'vs/workbench/common/composite';
 import { RawContextKey } from 'vs/platform/contextkey/common/contextkey';
+import { IPaneComposite } from 'vs/workbench/common/panecomposite';
 
 export const SideBarVisibleContext = new RawContextKey<boolean>('sideBarVisible', false);
 export const SidebarFocusContext = new RawContextKey<boolean>('sideBarFocus', false);
 export const ActiveViewletContext = new RawContextKey<string>('activeViewlet', '');
 
-export interface IViewlet extends IComposite {
+export interface IViewlet extends IPaneComposite {
 
 	/**
 	 * Returns the minimal width needed to avoid any content horizontal truncation
