+++ src/vs/workbench/contrib/experiments/browser/experimentalPrompt.ts
@@ -7,7 +7,7 @@ import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { INotificationService, Severity, IPromptChoice } from 'vs/platform/notification/common/notification';
 import { IExperimentService, IExperiment, ExperimentActionType, IExperimentActionPromptProperties, IExperimentActionPromptCommand, ExperimentState } from 'vs/workbench/contrib/experiments/common/experimentService';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
-import { IExtensionsViewlet } from 'vs/workbench/contrib/extensions/common/extensions';
+import { IExtensionsViewPaneContainer } from 'vs/workbench/contrib/extensions/common/extensions';
 import { IWorkbenchContribution } from 'vs/workbench/common/contributions';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { language } from 'vs/base/common/platform';
@@ -71,7 +71,7 @@ export class ExperimentalPrompts extends Disposable implements IWorkbenchContrib
 						this.openerService.open(URI.parse(command.externalLink));
 					} else if (command.curatedExtensionsKey && Array.isArray(command.curatedExtensionsList)) {
 						this.viewletService.openViewlet('workbench.view.extensions', true)
-							.then(viewlet => viewlet as IExtensionsViewlet)
+							.then(viewlet => viewlet?.getViewPaneContainer() as IExtensionsViewPaneContainer)
 							.then(viewlet => {
 								if (viewlet) {
 									viewlet.search('curated:' + command.curatedExtensionsKey);
