+++ src/vs/workbench/contrib/format/browser/showExtensionQuery.ts
@@ -4,12 +4,12 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
-import { VIEWLET_ID, IExtensionsViewlet } from 'vs/workbench/contrib/extensions/common/extensions';
+import { VIEWLET_ID, IExtensionsViewPaneContainer } from 'vs/workbench/contrib/extensions/common/extensions';
 
 export function showExtensionQuery(viewletService: IViewletService, query: string) {
 	return viewletService.openViewlet(VIEWLET_ID, true).then(viewlet => {
 		if (viewlet) {
-			(viewlet as IExtensionsViewlet).search(query);
+			(viewlet?.getViewPaneContainer() as IExtensionsViewPaneContainer).search(query);
 		}
 	});
 }
