+++ src/vs/workbench/contrib/search/browser/search.contribution.ts
@@ -52,10 +52,10 @@ import { ISearchHistoryService, SearchHistoryService } from 'vs/workbench/contri
 import { FileMatchOrMatch, ISearchWorkbenchService, RenderableMatch, SearchWorkbenchService, FileMatch, Match, FolderMatch } from 'vs/workbench/contrib/search/common/searchModel';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IPanelService } from 'vs/workbench/services/panel/common/panelService';
-import { ISearchConfiguration, ISearchConfigurationProperties, PANEL_ID, VIEWLET_ID, VIEW_CONTAINER, VIEW_ID } from 'vs/workbench/services/search/common/search';
+import { ISearchConfiguration, ISearchConfigurationProperties, PANEL_ID, VIEWLET_ID, VIEW_ID, VIEW_CONTAINER } from 'vs/workbench/services/search/common/search';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
-import { ExplorerViewlet } from 'vs/workbench/contrib/files/browser/explorerViewlet';
+import { ExplorerViewPaneContainer } from 'vs/workbench/contrib/files/browser/explorerViewlet';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
 
 registerSingleton(ISearchWorkbenchService, SearchWorkbenchService, true);
@@ -320,9 +320,11 @@ CommandsRegistry.registerCommand({
 		const contextService = accessor.get(IWorkspaceContextService);
 		const uri = fileMatch.resource;
 
-		viewletService.openViewlet(VIEWLET_ID_FILES, false).then((viewlet: ExplorerViewlet) => {
+		viewletService.openViewlet(VIEWLET_ID_FILES, false).then((viewlet) => {
+			const explorerViewContainer = viewlet?.getViewPaneContainer() as ExplorerViewPaneContainer;
+
 			if (uri && contextService.isInsideWorkspace(uri)) {
-				const explorerView = viewlet.getExplorerView();
+				const explorerView = explorerViewContainer?.getExplorerView();
 				if (explorerView) {
 					explorerView.setExpanded(true);
 					explorerService.select(uri, true).then(() => explorerView.focus(), onUnexpectedError);
